[Enhance][ Separate instructions & parameters in build Feature ]
  1. Separate instructions from parameters and let parameters can set for
     different command.
  2. Add delay to make sure if user is in powershell, "cmd" command can
     finish it's job before other orders are issued.

+++ src/M01_MainEditorFunction.ts
@@ -10,15 +10,16 @@ const Buildlog      = WorkSpace + "BuildLog.log";
 const EnvCheck      = WorkSpace + ".vscode/EnvCheck";
 var   BuildPath     = "";
 var   PreBuildCmd   = "&";
-var   BuildCmd01    = "";
-var   BuildCmd02    = "";
-var   BuildCmd03    = "";
+var   BuildCommand  = "";
+var   Parameter01   = "";
+var   Parameter02   = "";
 var   CleanCommand  = "";
+var   BuildStatus   = 0;
 //
 //  Change encode into 437
 //    cmd /C to compatible between MS cmd and MS Powershell
 //
-var GlobalCommand  = "cmd /C \"chcp 437 & cd " + BuildPath + " & ";
+var GlobalCommand  = "cmd /C \"cmd & chcp 437 & cd " + BuildPath + " & ";
 
 //
 // Clear Env file when project start.
@@ -32,21 +33,36 @@ FileSys.unlink (EnvCheck,(_err)=>{});
 //
 function GetTerminal (Message:string) {
     //
+    // Get Config value.
+    //
+    let GetConfig = vscode.workspace.getConfiguration();
+    //
+    // Check build path is exist or not.
+    //
+    BuildPath     = (GetConfig.get("CAT.00_BuildPath")+"").replace(/\\/g, "/").indexOf (":/") === -1?
+                    WorkSpace + GetConfig.get("CAT.00_BuildPath") : GetConfig.get("CAT.00_BuildPath")+"";
+    if (!FileSys.existsSync(BuildPath) && !FileSys.existsSync(WorkSpace+BuildPath)) {
+        vscode.window.showInformationMessage (" ❗️❗️ File path  ["+BuildPath+"]  seems not exist.");
+        return null;
+    }
+    //
     // Reacquire all varlable, because user may change it any time.
     //
-    BuildPath     = (vscode.workspace.getConfiguration().get("CAT.BuildPath")+"").replace(/\\/g, "/").indexOf (":/") === -1?
-                    WorkSpace + vscode.workspace.getConfiguration().get("CAT.BuildPath"):
-                    vscode.workspace.getConfiguration().get("CAT.BuildPath")+"";
-    PreBuildCmd   = vscode.workspace.getConfiguration().get("CAT.PreBuildCmd")+"&" !== PreBuildCmd ?
-                    vscode.workspace.getConfiguration().get("CAT.PreBuildCmd")+"&"+ DelEnvCheck() :
-                    PreBuildCmd;
-    BuildCmd01    = vscode.workspace.getConfiguration().get("CAT.BuildCmd01") !== "" ?
-                    "(" + PreBuildCmd + vscode.workspace.getConfiguration().get("CAT.BuildCmd01") + ") > "+ Buildlog + " 2>&1" : "";
-    BuildCmd02    = vscode.workspace.getConfiguration().get("CAT.BuildCmd02") !== "" ?
-                    "(" + PreBuildCmd + vscode.workspace.getConfiguration().get("CAT.BuildCmd02") + ") > "+ Buildlog + " 2>&1" : "";
-    BuildCmd03    = vscode.workspace.getConfiguration().get("CAT.BuildCmd03") !== "" ?
-                    "(" + PreBuildCmd + vscode.workspace.getConfiguration().get("CAT.BuildCmd03") + ") > "+ Buildlog + " 2>&1" : "";
-    CleanCommand  = "" + vscode.workspace.getConfiguration().get("CAT.CleanCmd");
+    PreBuildCmd   = GetConfig.get("CAT.01_PreBuildCmd")+"&" !== PreBuildCmd ?
+                    GetConfig.get("CAT.01_PreBuildCmd")+"&"+ DelEnvCheck() : PreBuildCmd;
+    BuildCommand  = GetConfig.get("CAT.02_BuildCmd") !== ""  ? "(" + PreBuildCmd + GetConfig.get("CAT.02_BuildCmd") + ") > "+ Buildlog + " 2>&1" : "";
+    if (GetConfig.get("CAT.04_SetParameterWith") === "Build") {
+        Parameter01   = GetConfig.get("CAT.Parameter01") !== "" ?
+                        "(" + PreBuildCmd + GetConfig.get("CAT.02_BuildCmd") + " " +GetConfig.get("CAT.Parameter01") + ") > "+ Buildlog + " 2>&1" : "";
+        Parameter02   = GetConfig.get("CAT.Parameter02") !== "" ?
+                        "(" + PreBuildCmd + GetConfig.get("CAT.02_BuildCmd") + " " +GetConfig.get("CAT.Parameter02") + ") > "+ Buildlog + " 2>&1" : "";
+    } else {
+        Parameter01   = GetConfig.get("CAT.Parameter01") !== "" ?
+                        "(" + PreBuildCmd .replace ("&", " "+GetConfig.get("CAT.Parameter01")+"&"+ GetConfig.get("CAT.02_BuildCmd")) + ") > "+ Buildlog + " 2>&1" : "";
+        Parameter02   = GetConfig.get("CAT.Parameter02") !== "" ?
+                        "(" + PreBuildCmd .replace ("&", " "+GetConfig.get("CAT.Parameter02")+"&"+ GetConfig.get("CAT.02_BuildCmd")) + ") > "+ Buildlog + " 2>&1" : "";
+    }
+    CleanCommand  = "" + GetConfig.get("CAT.03_CleanCmd");
     GlobalCommand = "cmd /C \"chcp 437 & cd " + BuildPath + " & ";
     //
     // Create Terminal.
@@ -147,41 +163,44 @@ function Delay (Sec :number){
 //
 export async function CreatEnvAndBuildCode () {
 
-    const Terminal  =  GetTerminal (' 🔍 Checking build environment..... ');
+    const Terminal  = GetTerminal (' 🔍 Checking build environment..... ');
+    var   GetConfig = vscode.workspace.getConfiguration();
+    if (Terminal === null) { return; }
     //
     // Check Build commands.
     //
-    if (BuildCmd01 === "") { 
-        vscode.window.showInformationMessage (' ❗️❗️ Please unless set [CAT.BuildCmd01]');
+    if (PreBuildCmd === "&") { 
+        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.01_PreBuildCmd] before you use this function.');
         return;
-    } else if (PreBuildCmd === "&") { 
-        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.PreBuildCmd] before you use this function.');
+    }else if (BuildCommand === "") { 
+        vscode.window.showInformationMessage (' ❗️❗️ Please unless set [CAT.02_BuildCmd]');
         return;
     }
     //
     // Check if there have 02 or 03 command.
     //
-    if (BuildCmd02 !== "" || BuildCmd03 !== "") {
+    if (Parameter01 !== "" || Parameter02 !== "") {
         await vscode.window.showInformationMessage (
-            " 🤔 Chouse time !! select the command that you want to execute",
-            "1⃣️  "+vscode.workspace.getConfiguration().get("CAT.BuildCmd01"),
-            "2⃣️  "+vscode.workspace.getConfiguration().get("CAT.BuildCmd02"),
-            "3⃣️  "+vscode.workspace.getConfiguration().get("CAT.BuildCmd03"))
+            " 🤔 Chouse time !! select one command to execute ~~",
+            "1⃣️  Build without parameter.",
+            "2⃣️  With ["+GetConfig.get("CAT.Parameter01")+"]",
+            "3⃣️  With ["+GetConfig.get("CAT.Parameter02")+"]")
         .then (function (Select) {
             let Build = Select?.indexOf("1⃣️  ") !== -1 ? 
-                        BuildCmd01 : Select?.indexOf("2⃣️  ") !== -1 ?
-                        BuildCmd02 : BuildCmd03;
+                        BuildCommand : Select?.indexOf("2⃣️  ") !== -1 ?
+                        Parameter01  : Parameter02;
             if (!Select || Build === "") {
                 vscode.window.showInformationMessage (' ❗️❗️ Cancel execution.');
                 return;
             }
             Terminal.sendText (GlobalCommand + Build + "\"");
+            vscode.window.showInformationMessage (" 🐈 Start to build code.");
         });
     } else {
-        Terminal.sendText (GlobalCommand + BuildCmd01 + "\"");
-        await Delay(100);
+        await Delay(1000);
+        Terminal.sendText (GlobalCommand + BuildCommand + "\"");
+        vscode.window.showInformationMessage (" 🐈 Start to build code.");
     }
-    vscode.window.showInformationMessage (" 🐈 Start to build code.");
     Terminal.show (true);
     const options = {
         selection: new vscode.Range (new vscode.Position(0, 0), new vscode.Position(0, 0)),
@@ -194,21 +213,22 @@ export async function CreatEnvAndBuildCode () {
 //
 //  Clean up work space
 //
-export function CleanUpWorkSpace () {
+export async function CleanUpWorkSpace () {
 
     const Terminal  =  GetTerminal (" 🧹 Start to clean up your work spase.");
+    if (Terminal === null) { return; }
     //
     // Check Clena command.
     //
     if (CleanCommand === "") { 
-        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.CleanCmd] before you use it.');
+        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.03_CleanCmd] before you use it.');
         return;
-    }
+    } await Delay(1000);
     //
     // Delete Build log and clean workspace.
     //
     FileSys.unlink (Buildlog,(_err)=>{});
-    Terminal.sendText (GlobalCommand + CleanCommand + "\"");
+    Terminal.sendText (GlobalCommand + PreBuildCmd + CleanCommand + "\"");
     Terminal.show (true);
 }
 
@@ -263,20 +283,21 @@ export function ChecBuildLogAndJump2Error () {
 export async function BuildSingleModule () {
 
     const Terminal  =  GetTerminal (' 🔍 Checking build environment..... ');
+    if (Terminal === null) { return; }
     //
     // Check prebuild command.
     //
     if (PreBuildCmd === "&") { 
-        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.PreBuildCmd] before you use this function.');
+        vscode.window.showInformationMessage (' ❗️❗️ Please set [CAT.01_PreBuildCmd] before you use this function.');
         return;
-    }
+    } await Delay(1000);
     //
     // Check terminal environment.
     //
     if (!FileSys.existsSync(EnvCheck)) {
         vscode.window.showInformationMessage (' 🔬 Check prebuild command can work or not.... ');
         Terminal.sendText (GlobalCommand + "("+ PreBuildCmd +"nmake -t > "+ EnvCheck + " 2>&1)" + "\"");
-        await Delay(100);
+        await Delay(500);
         do {/* Wait for EnvCheck create */} while (!FileSys.existsSync(EnvCheck));
     }
     let CheckFile = FileSys.readFileSync (EnvCheck, 'utf-8');
@@ -297,7 +318,6 @@ export async function BuildSingleModule () {
     //
     //  Start find current file in inf and module path in build folder.
     //
-    await Delay(100);
     let FileName = vscode.window.activeTextEditor?.document.fileName.replace(/\\/g,"/").split("/").pop()+"";
     let ModuleName = FindModuleName (vscode.window.activeTextEditor?.document.fileName+"", FileName, ".inf", BuilFolder);
     if (ModuleName === "") {
@@ -309,7 +329,7 @@ export async function BuildSingleModule () {
         vscode.window.showInformationMessage (" ❗️❗️ Can\'t find [ "+ModuleName+" ] in Build folder, please unliess full build once time or make sure this module will been build.");
         return;
     }
-    vscode.window.showInformationMessage (' 🐈 Start to build module [ '+ModuleName+' ].');
+    vscode.window.showInformationMessage (' 🐈 Start to build module =>[ '+ModuleName+' ].');
     let ModuleBuildCmd = "";
     for (let i=0; i<MakeFilePath.length; i++) {
         ModuleBuildCmd = ModuleBuildCmd+ PreBuildCmd +"nmake -f "+MakeFilePath[i]+"&";
